[CI] Add initial CI
This PR adds two simple GitHub Actions Workflows:
* builds.yml: Which verifies that our commits and pull requests properly build and run through tests
* lint.yml: Which lints only changed lines of code in a PR or commit for either linting issues or code style/formatting violations

Part of the complexity is installing all the dependencies. The action.yml in dependencies installs basically all readily available dependencies that we need (Python, cmake, JDK, Compilers, ccache).

The action.yml in llvm-build does a build of the LLVM 16 release branch and caches it (as it otherwise takes roughly 2 hours to build).
The advantage of building our own LLVM are that we can use better debugging options such as sanitizers and are not reliant on any kind of repository but the source repository.

While doing issues popped up in regards to both sanitizer issues and macOS issues. I have also fixed these as part of this PR. I can separate them in a different PR if you prefer but I feel like without test infrastructure they're somewhat not meaningful