[object] Add structs and classes for representing in memory layout of Java objects
The main use for them right now is mostly for offset and size calculations within the VM. We previously had hardcoded constant or odd looking seemingly unrelated `sizeof` expressions. With the new code we have static functions to get various byte offsets and we can use `sizeof` expressions on actual classes.

This will also give us greater flexibility in the future if we ever change anything about the layout of these objects.