[classloader] Eagerly create array types if possible
Creating the class object for an array type has no visible side effects unlike creating class objects of normal class objects (which can cause exceptions and static initializers to run etc).
It is therefore a useful optimization to create array objects eagerly if possible. This is the case when the component type of the array has already been loaded.

Creating class objects eagerly leads to better code generation in the JVM Bytecode to LLVM IR compilation and less stubs and trampolines being created.