[llvm] Fix crash due to a wrong function attribute on our allocation function (#12)
The `noalias` attribute on return types of functions has additional
special semantics in LLVM. According to the LangRef, it makes LLVM
assume that the pointer returned by the function is unique AND that it
comes from the system allocator directly.

The latter has bad consequences in our case however. It makes LLVM
assume that stores to the objects memory that are not read are dead and
can be removed. This is incorrect in our case however as the GC relies
on code setting the class object and attempts reading it. LLVM deleting
the store of the class object however leads to it remaining null, and
the GC crashing when trying to parse the heap during garbage collection.

This PR fixes that issue by removing the inaccurate `noalias` attribute
and replacing it with the `allockind` attribute
(https://llvm.org/docs/LangRef.html#function-attributes) this makes LLVM
assume that any stores to the object might be read by the allocation
function while still assuming that any pointers returned by the function
may not alias any other objects.