[vm] Add a high level API to easily implement 'native' methods of Java in C++
The high level API builds on top of the JNI interface, simply doing casts to go from the JVM agnostic to classes specific to our JVM implementation.
The API essentially works by creating 'Model' classes in C++ which should mirror the class from Java. Within the 'Model' class, one defines (possibly static) methods with the exact same name and signatures (translated to C++ types) as are used in Java. One then only has to list the name of the Java class and the list of methods implemented in static members of the model class and register in a 'VirtualMachine' instance.

As a simple demonstration of how to use the API, a model class for Javas 'Object' was implemented with an implementation of 'hashCode'. It lazily computes a unique id for a Java object and then stores it within the object header to remain stable throughout the objects lifetime. This is similar to how Chromium's V8 JS engine does it as well. Using the objects address is not viable due to being unstable within our relocating Garbage Collector.