[codegen] use `zext` instead of `sext` when working with `char`
The original code incorrectly assumed that `char` was a signed type. The JVM spec specifies it as (the only!) unsigned type however, so doing a signed extension leads to correct and unexpected results in various places.

This PR fixes this issue in various places by checking the descriptor whether the integer type is signed or unsigned and using proper instructions accordingly.

Fixes https://github.com/JLLVM/JLLVM/issues/44