[CI] Update `clang-format-diff.py` from upstream (#62)
In https://github.com/JLLVM/JLLVM/pull/59 we encountered issues with the
script failing to filter files diffs that have only removals in their
changes. This has fixed upstream in
https://github.com/llvm/llvm-project/commit/50563944ab962b58a1e00763ce16d8c712965c6d
but is not yet part of any LLVM release.
Luckily this is only a simple single-file python script that we can
import and reuse.

This PR does exactly that.