[object][instruction] Make `ClassObject` a proper Java object (#75)
`Class` is Java's most fundamental mechanism for reflection and very
important to its execution model. We additionally utilize it as RTTI
object and owner of V and ITables. These have so far been in an internal
format however which was incompatible with the `Class` shipped with
OpenJDK.

This patch fixes that by making it a proper java object and
synchronizing its field layout with OpenJDK. The largest complexity in
the patch is actually a change to the class loader. Since `Class` is
also its own class object, a special "bootstrap load sequence" has to be
done in the class loader on startup. Some of the fields from OpenJDK are
likely to require special initialization by the VM that is currently not
yet done and left as future exercise once required.

Additionally, because testing would be difficult without it, `ldc` for
class objects was implemented and a few of the `native` methods of
`Class` and `Object` for very basic testing.

It is expected that more tests with finer grained testing of various
methods and aspects in `Class` will be added in the future on a per need
basis.

Fixes https://github.com/JLLVM/JLLVM/issues/7 and
https://github.com/JLLVM/JLLVM/issues/8