[codegen] add LLVM module verification in debug builds and fix discovered issues
LLVM by default does not a lot of structural non-local verifications, even in debug builds of LLVM. It instead requires an explicit call to a verify function which then outputs proper error messages.

Not having the verifier turned on has actually lead to silent bugs in our codegen which have luckily not really manifested themselves up until now but would lead to broken preconditions in the optimizer (I have another patch coming up where it does result in a crash).
The issue was simply that `sext` and `zext` attributes were accidentally being applied to the wrong parameters in non-static methods due to not taking the `this` parameter into account.
This patch fixes that.