[codegen] fix false positives when creating basic blocks (#84)
This is achieved by creating a new byte code iterator abstraction which
centralizes our logic for iterating over JVM bytecode, allowing reuse of
the logic both when creating the basic blocks and generating LLVM IR.

Previously the basic block creation would interpret everything that
looks like a branch instruction as a branch instruction. This would
create incorrect basic blocks with no terminators which would make the
verifier in LLVM fail as well as some optimization passes.

The implementation was intentionally kept very simple and a large focus
was made on having to modify as little code as possible, keeping the
code for all instructions implemented so far identical. One could
definitely build a nicer abstraction with sum types that already contain
parsed data for the instruction but this would require larger code
changes that I don't think are worth my time right now.