[vm] Add command line option to conveniently enable `LLVM_DEBUG` logging
Output of `LLVM_DEBUG` can often be used for debugging purposes, especially when trying to trace actions. Debugging linking failures with `-Xdebug=jitlink` or finding out which method was last compiled prior to a crash using `-Xdebug` (equivalent to `-Xdebug=jvm`) are very useful to quickly locate an issue. This commit adds these command line arguments and enables them in debug builds.

The reason why only in debug builds is that these debugging facilities are not compiled into release builds.