[instruction][fix] Implement `ireturn` for functions returning smaller integer types (#89)
When returning integers in functions returning `i16` or smaller we have
to truncate the `i32` first to not create an invalid LLVM IR module that
does not verify due to `ret` type mismatches. Additionally, booleans
have to be treated specially, as the spec states that converting to a
boolean return type has to be done via a bitmask with 1.