[support][cmake] Enable C++20 and import better 3rd party variant (#91)
Motivation is the code in https://github.com/JLLVM/JLLVM/pull/86

I believe its much nicer but it has the issue of not great compile times
which are especially an issue for clang-tidy for some reason? I found
https://github.com/groundswellaudio/swl-variant instead which requires
C++20 however.

We were so far on C++17 mostly because LLVM is built this way and also
trying to figure out what compilers support what is possible painful.
Part of the purpose of this PR is to figure out how painful... We use
very recent GCC and Clang versions on Linux, but I have absolutely no
clue about AppleClang.

Lets just enable C++20 for now and carefully use features. We luckily
have CIs to make sure our code compiles on platforms we currently care
about.