[fix] properly take superinterfaces of superclasses into account as well
The iterators used to enumerate all interfaces (either in any order or topological) previously only took into account direct superinterfaces and their transitive superinterfaces but not the (direct and indirect) superinterfaces of superclasses.
This would lead to failures to allocate and populate ITables for specific classes followed by runtime crashes when `invokeinterface` did not find an ITable entry.

This patch fixes that by changing the GraphTraits to create a whole classgraph out of all class objects and simply use a filter to only output interfaces. This makes the superinterfaces of superclasses reachable as well.

The implementation of ClassObject was also adjusted to ease the implementation.