[vm] implement basic implementation for 'assert' class initializers
Any uses of `assert` within a class leads to the creation of a class initializer which queries the assertion status from the class loader and VM: https://godbolt.org/z/9a6q675bx

This PR implements the 'native' method called from `Class.desiredAssertionStatus` to allow our VM to load class files that contain `assert` (even if they're never executed). `assert`s themselves do not yet work (nor can be compiled) since it contains `athrow` of `AssertionError`.

I currently made it simply match whether the VM is compiled in debug or release but this isn't really proper. Rather, a command line option should be added in the future to make this user configurable, but this is beyond the scope of this PR