[instruction] add the most basic implementation of `monitorenter`, `monitorexit` and `athrow`
Multi threading is currently a non-goal for us, therefore this PR currently implements them as doing nothing but the require operand stack manipulation.
We might still be executing and compiling `synchronized` blocks however, hence their implementations are required.

A `athrow` stub is required due to `synchronized` catching any exception and then rethrowing it. This can currently never be executed, so the stub that does the right stack manipulation should suffice for now