[codegen] implement basic compilation of catch sections
While these are currently not reachable (since we do not yet have an exception mechanism), catch handlers are part of the bytecode stream and therefore have to be compiled by our VM.

Catch handlers have the special semantic of their operand stack having only the exception on the very top which we have to register within our code. Otherwise compilation would lead to crashes and out of bound reads.