[object] make vtable slot assignment correct first, optimized later
The current vtable slot assignment was prematurely optimized to the point of being incorrect. In particular, calling private methods was not possible as they have been filtered out. In the future, if we were to fully implement the complicated package private overriding scenario described here https://docs.oracle.com/javase/specs/jvms/se17/html/jvms-5.html#jvms-5.4.5 it'd likely stand in our way too.

For now, just simplify the implementation and give a vtable slot to every non-static method. Then properly assign method according to the method selection described in 5.4.6 (this is the same procedure as for interface methods).
Additionally, package private calls from the same "package" was implemented (this is required to load `Class`) and tests annotated XFAIL with todos for the future were added