[instruction] Implement `athrow` and unwinding mechanism (#103)
This if the first step for exception handling in our JVM. As described
in https://github.com/JLLVM/JLLVM/issues/6, due to shortcomings in LLVMs
builtin exception handling when using a relocating GC, we are opting to
build the mechanism ourselves. This is achieved by manually generating
code to check if and exception was thrown and do appropriate dispatch.

This patch is the first patch in this series which implements `athrow`
and the unwinding mechanism. Throwing essentially sets a global called
`activeException` to the now active exception. Unwinding is done by
checking whether this global is set and returning from the function if
true. This requires generating code after every instruction might cause
unwinding.

Next patches will create code for dispatching into catch handlers.

Fixes https://github.com/JLLVM/JLLVM/issues/54