[GC][codegen] add extra GC error detection using ASAN (#107)
One of the hardest to debug errors once they occur are errors related to
GC. Most of these are usually a relocation not being updated or maybe
the GC collecting an object to early.

Either way, the earlier and faster we can detect these errors the
better. This patch adds support for using ASAN to instrument the
`toSpace` aka the not actively in use heap area by poisoning it to make
any reads and writes instrumented with ASAN fail with a backtrace and
error message. To make this more even more effective, passes from LLVM
are added to our JIT to also instrument our produced Java code.

This makes it possible to also detect any incorrect memory operations in
Java due to codegen errors and similar.