[refactor] Consolidate state from `codeGenBody` into struct
I think we can all agree that `codeGenBody` has gotten a bit large. The `match` statement itself isn't that horrible IMO but rather a lot of the state and variables floating around which has lead to the creation of lambdas and inline code which should have really been functions but weren't out of laziness of not wanting to pass dozens of parameters around.

This patch therefore focus on fixing that part of the equation by encapsulating all things related to codegen within a class, making it possible to access all state relevant for the generation of code from within any methods of that struct.
This most importantly makes it easier to extract code from e.g. `match` statements to allow code reuse.

Some functions, like the calculation of basic blocks, or the lambdas for EH generation have already been extracted to methods.

Only admittedly odd thing is all the work happening in the constructor... Anything else felt overkill to me.