[test] add Catch2 as unit testing library (#113)
We are currently only relying on integration tests for our test suite
which nicely ensures that all components are working but has the issue
of not really properly allowing to test small details/aspects within our
components.

I'd therefore like to add Catch2 as a unit testing library. My concrete
use case is that I want to test various allocation patterns in the GC
for correctness in an upcoming revision.

To conveniently add the library everywhere without having to vendor it,
CPM was used which automatically downloads the library, includes it in
cmake and makes it available. Additionally, Catch2 has great integration
with test discovery with `ctest` (the testing driver we use), making it
run all `TEST_CASE`s in a single process in parallel

A `HelloWorldTest` was added as "demonstration/example" and should be
removed once a proper unit test is added