[GC] add concept of GC roots which can be used from C++ code to persist Java object references (#114)
A plain `T*` is currently the only way to refer to Java objects from C++
but has the large disadvantage of not persisting passed garbage
collections. Instead, the pointer gets left dangling either due to the
GC having deleted the object, as the C++ pointer is not visible to the
GC, or because the object has been relocated and the GC cannot update
the C++ pointer.

For that reason the concept of GC roots has been added which are used in
C++ code to refer to Java objects without any of the disadvantages of a
`T*` with the slight cost of runtime overhead (obviously correctness is
more important than performance here).

A special allocator called `RootFreeList` has been added to allocate
these roots and the GarbageCollector extended to be able to use these. A
convenient class owning an instance of a root has been added to
automatically deallocate a root with RAII once no longer required and a
lightweight class was added to refer to a root with the roots lifetime
managed elsewhere.

Unit tests have also been added for all these features and convenient
`allocate` methods for objects and arrays added to the GC.

Most importantly, a small user guide when to use `T*`, `GCUniqueRoot` or
`GCRootRef*` was added to the description of the `GarbageCollector`.