[JNI][NFC] Exclusively use GC roots in the JNI interface
The JNI spec requires `jobject` (which in our implementation will be an alias to `GCRootRef`) to be rooted.
This PR therefore adjusts the bridge code generation to automatically create and delete these roots. Since our higher level interface
in NativeImplementation.hpp builds on top of the JNI, it had to also be adjusted to now work on
GCRootRef instead of raw pointers.
Documentation and implementations were therefore adjusted accordingly.