[refactor][NFC] Move implementations of internal functions into its own Dylib (#116)
These have so far been directly implemented in the main Dylib.
Implementing them in a separate dylib allows reuse of that dylib in the
link order of other dylibs. Specifically, I'd like to reuse the
implementation of these internal symbols in the JNI bridge dylib as
well.

Additionally, a nice API was added to add these methods and the
implementations were moved to the VirtualMachine. The reason for the
latter is that the implementations of these functions should have more
access to all the required state (including active exception, class
loader, etc) available from the Virtual machine without having to
needlessly pass them into the JIT instance.