[vm] eagerly load string class object for string interner (#118)
We cannot load the string class object lazily due to it occurring during
bytecode compilation. Bytecode compilation, due to being lazy/on-demand
is implemented using trampolines in the JIT which do a kind of
"context-switch", making the Java stack unreachable for the unwinder and
causing the GC to think all objects so far are unreachable. This would
cause all of them to be deleted causing massive issues.

This patch fixes that by simply eagerly loading the string object for
the string interner.