[vm] stop running class initializers during bytecode compilation
As we recently discovered, due to the context switching of the JITs trampoline, the Java stack becomes unreachable to the GC while we are compiling bytecode. If a GC then occurs due to running Java code, the GC deletes essentially all objects on the Heap.

It is therefore required that Java code is NOT executed during compilation. Prior to this PR, the class loader would initialize class objects during compilation of stubs (whose whole responsibility was the loading of classes) which would then execute Java code if the class had a class initializer.

This PR fixes that by reworking how classes are initialized.
First, the responsibility is taken away from the class loader as this was simply the incorrect place to do so. According to the spec, classes are only initialized by specific instructions with only a few exceptions.
Therefore initialization has now been implemented in the VirtualMachine.
Next, if an instruction must initialize the class object (these are `invokestatic`, `new`, `getstatic` and `putstatic`), the call to the initialization routine is now compiled into the stub if required so that it is executed on the Java stack.

See the spec here https://docs.oracle.com/javase/specs/jvms/se17/html/jvms-5.html#jvms-5.5