[codegen] Properly implement `invokeinterface` and `invokevirtual` to spec
These two instructions are very special in that each of the two may act like the other!
Really, the only difference between the two is how the precise methods are resolved from the specified class, method and type names.
Besides that, each may act as calls into a vtable, calls into an itable or even direct calls in some scenarios.

For that reason, a common method was created to implement both of them at once. Since the information whether a vtable or itable call has to be made requires the class object to be loaded, the lazy class loader was used.
In the case that the class object is loaded, we essentially generate perfect inline code performing the required lookups.
In the case that the class object is not loaded during compilation, a small compile stub is created that is jumped to that then generates the actual vtable or itable lookups. This should be relatively cheap after compilation since the call is always a tail call and there are likely enough caller saved registers for the lookup operations.