[object][NFC] reduce number of VTable slots (#142)
Since our new `invokevirtual` implementation it is now perfectly
possible to create direct calls to methods if after resolution we know
they can't be overwritten. This is the case for any private or final
instance methods for example.

This PR uses that capability to reduce the amount of V-Table slots we
create for class objects.