[object][NFC] reduce number of Itable slots
Similar to https://github.com/JLLVM/JLLVM/pull/142, this reduces the amount of slots allocated in an I-Table for an interface.
Interface methods can't actually be final, hence no test was added for that case here.
Private interface methods can't be overwritten (same logic as for normal class methods) hence there is no need to create slots for them.