[codegen][NFC] add optimization attributes to builtins (#147)
Since we know the implementations of these functions, we can tell LLVM
some of their attributes in hopes of it generating better code.
The most important of the attributes added is `gc-leaf-function`. It
tells LLVM that the function can not cause garbage collection and that
LLVM does not have to generate code that makes the stack parseable at
that function call.
`nocapture` tells LLVM that the pointer passed does not escape in any
shape or form, which is mostly an aid for alias analysis.
Similarily, `memory`, here with "readOnly" and
"inaccessibleOrArgMemOnly" gives ensurance to LLVM that the function
calls does not modify any memory in its current scope.
`willreturn` and `nounwind` are very minor attributes that can enable
some code motion.