[codegen][gc] reliably retrieve RSP during unwinding
Unwinding the stack is a very complicated topic and it turns out DWARF CFI is very complicated.
When working with deeper call-stacks I have discovered multiple occurrences of the unwinder being unable to retrieve the stack pointer and crashing instead.

I have tried multiple combinations of: using libunwind instead of libgcc, all kinds of compiler flags, both in our JIT and C++ compiler, testing with GCC, using the level 1 libunwind API and so on and on, but would usually end up just having crashes elsewhere.

The version that basically worked perfectly for everything I've tested was fetching the Canonical Frame Address (CFA) which as far as I can tell is equal to the stack pointer but always available (https://maskray.me/blog/2020-11-08-stack-unwinding).
The code previously used `_Unwind_GetGR` which uses a dwarf register number and a more generic mechanism to be able to fetch any (callee-saved) register, but is seemingly unreliable in libgcc (?).

Future work could explore using libunwind and using its API for JIT frame registrations.