[gc] introspect into superclass fields as well
This is a very embarrassing oversight :) So far we've only been traversing into the direct fields of a class but forgot to traverse into the indirect fields of a class that it gained through inheritance.

This PR does a simple correct implementation for now to unblock work with a TODO how to properly optimize this in the future.