[codegen] also enable GC in generated stubs
The stubs might do garbage collection as part of executing a class initializer. Any arguments passed into the stub, that are then usually passed onto the actual caller function, have to both be 1) found by the GC and 2) relocated by the GC.
1) worked perfectly previously as the caller of the stub registers all arguments in the stackmap bug 2) did not.
This would cause a relocated object not be updated in the stack frame of the stub and a dangling pointer being passed to the actual caller.