[codegen] also enable GC in generated stubs (#162)
The stubs might do garbage collection as part of executing a class
initializer. Any arguments passed into the stub, that are then passed to
the actual callee, have to both be 1) found by the GC and 2) relocated
by the GC. 1) worked perfectly previously as the caller of the stub
registers all arguments in the stackmap bug 2) did not.
This would cause a relocated object to not be updated in the stack frame
of the stub and a dangling pointer being passed to the actual callee.