[native] implement `Reflection.getCallerClass` and callstack metadata (#159)
This method is used relatively extensively within the OpenJDK, mostly to
provide reflection capabilities. This includes at VM startup. The method
also served as a good motivation to implement adding metadata to
compiled Java methods that can be fetched during unwinding. This
mechanism allows getting reflective information about any Java method
from stacktraces. This mechanism can later also be used to implement
symbolized stacktraces of exceptions.

The basic mechanism used to implement all of this is 1) adding the
metadata into the `prefix` attribute of LLVM functions. These are placed
immediately before the function in the resulting assembly. 2) Marking
Java functions as originating from Java. This is achieved by putting
them into "java" sections. This is required for 3) maintaining a set of
all Java method addresses in the linker. This is used to identify a
method as having metadata.