[codegen] properly implement method resolution for `invokestatic` and `invokespecial` (#168)
The former is relatively simply and mostly requires method lookup for
the sake of finding static methods in superclasses.

`invokespecial` is a lot more complicated. It was actually previously
known as `invokenonvirtual` and would be used for invoking private
methods and `super` method calls as well. For the latter case it is
still used today as well, even within interfaces. Its semantics were
changed in later Java versions if the `ACC_SUPER` class is set,
requiring more lookups for resolving non-init methods. See
https://stackoverflow.com/questions/8949933/what-is-the-purpose-of-the-acc-super-access-flag-on-java-class-files

This PR implements these changes by closely following the spec:
https://docs.oracle.com/javase/specs/jvms/se17/html/jvms-6.html#jvms-6.5.invokespecial