[JNI] add concept of `local` frames to manage roots during JNI calls
We currently had issues avoiding double frees of roots when creating roots during JNI calls. The reason is that it is cumbersome to determine whether a root returned by a native function happens to be a root passed into the function or whether it is a global root and whatnot.

Instead, adopt a strategy that is actually mandated by the JNI spec: Create new "local root frames" in which all local roots during a JNI call are allocated that are then automatically freed on return. These also correspond to the `PushLocalFrame` and `PopLocalFrame` functions defined here https://docs.oracle.com/en/java/javase/17/docs/specs/jni/functions.html#pushlocalframe

With these, we can simply pop the frame at the end of a JNI call to free all local roots.