[object][native] implement unicode output in `System.out`
This mainly required telling the system via a native method whether we are big endian or not.
We apparently got the wrong idea that strings in Java are always big endian if UTF-16. Maybe this used to be true in older versions. Nowadays, they are apparently in native byte order and one simply has to tell the system the byte order.

With this patch we can now correctly display unicode characters in the output.