[vm] create internal `-Xno-system-init` option and use it in execution tests.
Our tests in `Execution` are currently pretty nice since they require very little other Java infrastructure and really only rely on the correct implementations of JVM Instructions.
Booting the whole VM with the system is therefore unnecessary and at this point in time heavily impacts our testing time.

This PR therefore adds `-Xno-system-init` which is also automatically turned on for any tests in the `Execution` directory in addition to the usual `-Xenable-test-utils`.
This would allow us to reuse the tests in `Execution` if we ever bring up alternative execution methods that aren't the LLVM JIT (e.g. interpreter, or another not-LLVM based JIT).

Any tests that rely on system initialization have instead been moved to the `System` directory.