[support] add `NonOwningFrozenSet`
This class allows creating a hashset over an array ref without taking ownership of the keys. The set is created and then frozen upon construction, making it not possible to add any more keys to the set.
The set additionally preserves insertion order.

One of the main features of the set is the support for heterogeneous lookup, making it possible to do lookups with any type with a compatible implementation to the key type.

Main use case for this set is in use of the 'ClassObject' class for field and method lookup. It is therefore also a standard layout type.