[docs] Add initial documentation (#183)
This PR is an initial draft at adding a deployable config for us to
write any project documentation.

The documentation system proposed is using Sphinx together with MyST.
The advantage of Sphinx is that it is extendable via plugins, has a rich
ecosystem and also produces static HTML files that can be viewed offline
without a running webserver.
The generation has been integrated with cmake and can be generated using
the `docs` target.
It copies and creates a `docs/build/` directory where the output
artifacts exist.

The advantage of MyST is that for 90% of tasks, one can simply use
normal markdown, which, in my opinion, is more comfortable than RST. Yet
when more advanced features of RST are required, it is possible to use
these as well using special syntax. See https://mystmd.org/ as
introduction

Last but not least, I used https://github.com/pradyunsg/furo as a theme
because I think it looks cool, has an automatic dark mode and a nice
TOC.

A GitHub pages workflow has been added for automatically deploying the
webpage as well.