[object][NFC] make `GCRootRef` implicitly convertible to `T*`
Previously, any API taking a `T*` would require explicitly casting a `GCRootRef`, either via `static_cast` or `.address()` to get the address of the contained object. This made using `GCRootRef` less comfortable then it has to be and made its behaviour differ to `Object*` from a user perspective in a few places.
Worst of all, wanting to avoid these explicit casts would often lead to APIs taking `GCRootRef<T>` parameters instead of `T*` parameters despite the latter being the better more generic choice if no GCs can occur during those function calls.

This PR therefore makes the conversion operator to 'T*' implicit. The advantages of this is the useability in all the previously mentioned scenarios.
The disadvantages are that one could deem it as non-obvious "magic" and the potential for confusing behaviour in templates (where the implicit conversion would not occur).
For the latter issue, the `JavaCompatible` and `JavaConvertible` concepts were added to make sure that a template type is of an expected type. `JavaConvertible` in particular can be used with the `JavaConvertedType` alias to then force required implicit conversions to `JavaCompatible` types as well.