[native] add `State` to `ModelBase` subclasses
By adding a type called `State` to subclasses, it becomes possible to persist data beyond a single function call of a `ModelBase` subclass. An instance of the given `State` class gets allocated when registered in the VM and then simply injected into the model instance if a member function or the static method if requested.

Prime use-case for this functionality is to save `InstanceFieldRef` or `StaticFieldRef` instances across function calls, initializing them once and then using them from then on. This saves on doing a field lookup each time.