[object][NFC] add `InstanceFieldRef` and `StaticFieldRef` concepts (#188)
These classes serve as strongly typed wrappers around a `Field*`
allowing more convenient access to a field from C++ code. Access is
modeled via the call operator, returning a reference to either the
static field or the instance field of an object.

The most important use-case for these is for native interop. Previously,
we were forced to create a C++ type mirroring the Java object layout to
be able to access any of its fields conveniently. This creates a tight
coupling between the C++ type and the corresponding class file which was
very fragile as the layout can easily change.

With these new classes, we can instead search for the field in question
and use it directly. The only coupling remaining are the field name and
type which are less fragile than the memory layout of the containing
object.

This has been done with the `Thread` class in this PR as a
demonstration.

Using a mirrored C++ type is still recommended in scenarios requiring
performance.

Depends on https://github.com/JLLVM/JLLVM/pull/187