[class] refactor descriptor classes (#190)
This PR is essentially a reimplementation of `FieldType`, `MethodType` &
friends.

These new implementations has the following goals:
* Compact representations: The new classes are 16 bytes each and trivial
classes. This makes it efficient to pass these by value.
* No memory allocation
* Standard layout type
* Composable: It is possible to create descriptors by hand by simply
constructing and passing the descriptor objects around. This gives us
correctness by construction unlike the string API
* constexpr friendly: If one does want to use the string representation
(for compactness e.g.), then using it with a string literal will do
`constexpr` construction and verification, leading to a compile time
error if the descriptor is invalid.

`FieldType` has essentially been implemented as a `variant`-like type
containing one of the three alternatives. The implementations of the
concrete alternative classes are implemented as wrappers around a
`FieldType`, making conversion to a `FieldType` trivial.

`MethodType` is mostly implemented with a single `string_view`-like
field. Iterating over its parameters is done with special iterators
passing a single `FieldType` at a time. An offset to the return type and
a parameter count are also stored.

This PR serves as the initial implementation. A follow up PR will update
the whole codebase to use these new descriptors for working with any
kind of descriptors instead of the manual strings we are currently
using.